requirements.txt
python-dotenv
pandas
langchain
langchain_openai
langchain_community
langchain-core
langchain-text-splitters
openai
numpy
faiss-cpu
flask

# 以下は、オプションで必要な場合に
# rank_bm25
# nltk